pyproject.toml
[project]
name = "necromancer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.5",
    "pygame-ce>=2.5.3",
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
python_files = "test_*.py"
